Initial implementation of custom split functions
It's possible to wire one up and get it called, not yet to do anything useful with it.
