+++ include/LightGBM/split_info.h
@@ -1,5 +1,5 @@
-#ifndef LIGHTGBM_TREELEARNER_SPLIT_INFO_HPP_
-#define LIGHTGBM_TREELEARNER_SPLIT_INFO_HPP_
+#ifndef LIGHTGBM_SPLIT_INFO_H_
+#define LIGHTGBM_SPLIT_INFO_H_
 
 #include <LightGBM/meta.h>
