+++ src/boosting/gbdt.h
@@ -111,6 +111,10 @@ class GBDT : public GBDTBase {
   */
   void ResetConfig(const Config* gbdt_config) override;
 
+  void SetSplitCallback(SplitCallback* callback) override {
+    tree_learner_->SetSplitCallback(callback);
+  }
+
   /*!
   * \brief Adding a validation dataset
   * \param valid_data Validation dataset
